% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_trip.R
\name{arcs_to_wkb}
\alias{arcs_to_wkb}
\alias{arcs_to_wkt}
\title{Convert arcs to WKT linestrings}
\usage{
arcs_to_wkb(x, ...)

arcs_to_wkt(x)
}
\arguments{
\item{x}{A wkpool (ideally after merge_coincident)}

\item{...}{Passed to \code{\link[wk:as_wkb]{wk::as_wkb()}}}
}
\value{
A wk_wkt vector of LINESTRING geometries
}
\description{
Convert arcs to WKT linestrings
}
\details{
Each arc (maximal segment sequence between nodes) becomes a linestring.
}
\examples{
x <- wk::as_wkb(c(
  "POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))",
  "POLYGON ((1 0, 2 0, 2 1, 1 1, 1 0))"
))
pool <- establish_topology(x)
merged <- merge_coincident(pool)
arcs_to_wkt(merged)

}
