% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_coincident.R
\name{as_decido}
\alias{as_decido}
\title{Convert wkpool to decido-ready format}
\usage{
as_decido(x)
}
\arguments{
\item{x}{A wkpool after merge_coincident}
}
\value{
A list with xy (vertex matrix) and segments (segment matrix, 0-indexed)
}
\description{
Convert wkpool to decido-ready format
}
\details{
Produces format suitable for decido::earcut() constrained triangulation.
Note: decido uses 0-indexed segments.
}
\examples{
x <- wk::as_wkb("POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))")
pool <- establish_topology(x)
as_decido(pool)

}
