% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_coincident.R
\name{find_neighbours}
\alias{find_neighbours}
\title{Build adjacency from shared edges}
\usage{
find_neighbours(x, type = c("edge", "vertex"))
}
\arguments{
\item{x}{A wkpool after merge_coincident}

\item{type}{"edge" for features sharing an edge, "vertex" for features sharing any vertex}
}
\value{
Data frame of feature pairs that are neighbours
}
\description{
Build adjacency from shared edges
}
