% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkpool.R
\name{plot.wkpool}
\alias{plot.wkpool}
\title{Plot a wkpool object}
\usage{
\method{plot}{wkpool}(x, col = NULL, ...)
}
\arguments{
\item{x}{A wkpool object.}

\item{col}{Colour(s) for segments. If \code{NULL} (default), segments are
coloured by feature using a built-in palette.}

\item{...}{Further arguments passed to \code{\link[=plot.default]{plot.default()}}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Draws segments coloured by feature membership.
}
\examples{
x <- wk::as_wkb(c(
  "POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))",
  "POLYGON ((1 0, 2 0, 2 1, 1 1, 1 0))"
))
pool <- establish_topology(x)
plot(pool)

}
