% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkpool.R
\name{pool_combine}
\alias{pool_combine}
\title{Combine wkpool objects}
\usage{
pool_combine(...)
}
\arguments{
\item{...}{wkpool objects to combine}
}
\value{
A single wkpool with merged pools and remapped segments
}
\description{
Combine wkpool objects
}
