% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_topology.R
\name{pool_compact}
\alias{pool_compact}
\title{Compact a pool by removing unreferenced vertices}
\usage{
pool_compact(x)
}
\arguments{
\item{x}{A wkpool}
}
\value{
A wkpool with only referenced vertices, .vx remapped
}
\description{
Compact a pool by removing unreferenced vertices
}
