% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_coincident.R
\name{topology_report}
\alias{topology_report}
\title{Report topology diagnostics}
\usage{
topology_report(x, tolerance = 1e-08)
}
\arguments{
\item{x}{A wkpool}

\item{tolerance}{Tolerance for "near miss" detection}
}
\value{
List of diagnostic information
}
\description{
Report topology diagnostics
}
\examples{
x <- wk::as_wkb(c(
  "POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))",
  "POLYGON ((1 0, 2 0, 2 1, 1 1, 1 0))"
))
pool <- establish_topology(x)
topology_report(pool)

}
