% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkpool.R
\name{vec_c}
\alias{vec_c}
\alias{vec_c.wkpool}
\title{Combine many wkpool vectors into one vector}
\usage{
\method{vec_c}{wkpool}(..., .ptype = NULL)
}
\arguments{
\item{...}{Vectors to coerce.}

\item{.ptype}{If \code{NULL}, the default, the output type is determined by
computing the common type across all elements of \code{...}.

Alternatively, you can supply \code{.ptype} to give the output known type.
If \code{getOption("vctrs.no_guessing")} is \code{TRUE} you must supply this value:
this is a convenient way to make production code demand fixed types.}
}
\value{
nothing, used for a message side-effect (see Details)
}
\description{
This is non-functional, wkpool does not currently support \code{\link[=vec_c]{vec_c()}}.
}
\details{
Attempts to combine wkpool vectors with vec_c will suggest
using [pool_combine()' instead.
}
\seealso{
\code{\link[=pool_combine]{pool_combine()}}
}
