% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_node.R
\name{arc_node_summary}
\alias{arc_node_summary}
\title{Summarize arc-node structure}
\usage{
arc_node_summary(x)
}
\arguments{
\item{x}{A wkpool (ideally after merge_coincident)}
}
\value{
List with counts and degree distribution
}
\description{
Summarize arc-node structure
}
\examples{
x <- wk::as_wkb(c(
  "POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))",
  "POLYGON ((1 0, 2 0, 2 1, 1 1, 1 0))"
))
pool <- establish_topology(x)
merged <- merge_coincident(pool)
arc_node_summary(merged)

}
