% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_node.R
\name{arc_node_summary}
\alias{arc_node_summary}
\title{Summarize arc-node structure}
\usage{
arc_node_summary(x)
}
\arguments{
\item{x}{A wkpool (ideally after merge_coincident)}
}
\value{
List with counts and degree distribution
}
\description{
Summarize arc-node structure
}
