% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_trip.R
\name{arcs_to_wkt}
\alias{arcs_to_wkt}
\alias{arcs_to_wkb}
\title{Convert arcs to WKT linestrings}
\usage{
arcs_to_wkt(x)

arcs_to_wkb(x, ...)
}
\arguments{
\item{x}{A wkpool}

\item{...}{Passed to wk::as_wkb}
}
\value{
A wk_wkt vector of LINESTRING geometries

A wk_wkb vector
}
\description{
Convert arcs to WKT linestrings

Convert wkpool to wkb
}
\details{
Each arc (maximal segment sequence between nodes) becomes a linestring.
}
