% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_node.R
\name{as_arcs}
\alias{as_arcs}
\title{Convert arcs to a wkpool of arc segments}
\usage{
as_arcs(x, arc_id = TRUE)
}
\arguments{
\item{x}{A wkpool (ideally after merge_coincident)}

\item{arc_id}{Logical: add .arc column to track arc membership?}
}
\value{
A wkpool with segments grouped by arc
}
\description{
Convert arcs to a wkpool of arc segments
}
\examples{
x <- wk::as_wkb(c(
  "POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))",
  "POLYGON ((1 0, 2 0, 2 1, 1 1, 1 0))"
))
pool <- establish_topology(x)
merged <- merge_coincident(pool)
as_arcs(merged)

}
