% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_node.R
\name{find_arcs}
\alias{find_arcs}
\title{Find arcs (maximal segment sequences between nodes)}
\usage{
find_arcs(x)
}
\arguments{
\item{x}{A wkpool (ideally after merge_coincident)}
}
\value{
A list of integer vectors, each containing .vx IDs forming an arc
}
\description{
Find arcs (maximal segment sequences between nodes)
}
\details{
Arcs are maximal paths through degree-2 vertices. They start and end
at nodes (degree != 2) or form closed loops through degree-2 vertices.
}
\examples{
x <- wk::as_wkb(c(
  "POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))",
  "POLYGON ((1 0, 2 0, 2 1, 1 1, 1 0))"
))
pool <- establish_topology(x)
merged <- merge_coincident(pool)
find_arcs(merged)

}
