% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_node.R
\name{find_nodes}
\alias{find_nodes}
\title{Find nodes (vertices where degree != 2)}
\usage{
find_nodes(x)
}
\arguments{
\item{x}{A wkpool (ideally after merge_coincident)}
}
\value{
Integer vector of .vx IDs that are nodes
}
\description{
Find nodes (vertices where degree != 2)
}
\details{
Nodes are branch points (degree 3+) or endpoints (degree 1).
Degree-2 vertices are pass-through points within an arc.
}
