% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkpool.R
\name{pool_combine}
\alias{pool_combine}
\title{Combine wkpool objects}
\usage{
pool_combine(...)
}
\arguments{
\item{...}{wkpool objects to combine}
}
\value{
A single wkpool with merged pools and remapped segments
}
\description{
Combine wkpool objects
}
\examples{
x <- wk::as_wkb("POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))")
y <- wk::as_wkb("POLYGON ((2 0, 3 0, 3 1, 2 1, 2 0))")
pool_a <- establish_topology(x)
pool_b <- establish_topology(y)
pool_combine(pool_a, pool_b)

}
