% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_trip.R
\name{segments_to_wkt}
\alias{segments_to_wkt}
\alias{segments_to_wkb}
\title{Convert wkpool segments to WKT}
\usage{
segments_to_wkt(x, type = c("multilinestring", "linestring", "point"))

segments_to_wkb(x, type = c("multilinestring", "linestring", "point"), ...)
}
\arguments{
\item{x}{A wkpool}

\item{type}{Output type: "linestring" (segments as paths), "multilinestring" (all segments),
or "point" (vertices only)}

\item{...}{Passed to \code{\link[wk:as_wkb]{wk::as_wkb()}}}
}
\value{
A wk_wkt vector
}
\description{
Convert wkpool segments to WKT
}
\examples{
x <- wk::as_wkb(c(
  "POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))",
  "POLYGON ((1 0, 2 0, 2 1, 1 1, 1 0))"
))
pool <- establish_topology(x)
segments_to_wkt(pool)
segments_to_wkt(pool, type = "linestring")

}
