% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_trip.R
\name{segments_to_wkt}
\alias{segments_to_wkt}
\alias{segments_to_wkb}
\title{Convert wkpool segments to WKT}
\usage{
segments_to_wkt(x, type = c("multilinestring", "linestring", "point"))

segments_to_wkb(x, type = c("multilinestring", "linestring", "point"), ...)
}
\arguments{
\item{x}{A wkpool}

\item{type}{Output type: "linestring" (segments as paths), "multilinestring" (all segments),
or "point" (vertices only)}
}
\value{
A wk_wkt vector
}
\description{
Convert wkpool segments to WKT
}
