% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc_node.R
\name{vertex_degree}
\alias{vertex_degree}
\title{Calculate vertex degree (number of segments touching each vertex)}
\usage{
vertex_degree(x)
}
\arguments{
\item{x}{A wkpool (ideally after merge_coincident)}
}
\value{
Named integer vector: names are .vx, values are degree
}
\description{
Calculate vertex degree (number of segments touching each vertex)
}
\examples{
x <- wk::as_wkb(c(
  "POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))",
  "POLYGON ((1 0, 2 0, 2 1, 1 1, 1 0))"
))
pool <- establish_topology(x)
merged <- merge_coincident(pool)
vertex_degree(merged)

}
