% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkpool-package.R
\docType{package}
\name{wkpool-package}
\alias{wkpool}
\alias{wkpool-package}
\title{wkpool: Vertex Pool Topology for Well-Known Geometry}
\description{
Establishes and maintains vertex pool topology for geometry handled by 'wk'. Segments are the atomic unit, vertices are shared via integer references into a pool. Topology is made discoverable via coincident vertex detection while not requiring modification of the input data. Topological data models follow principles described in Worboys and Duckham (2004, ISBN:978-0415283755). The edge-based topology geometry decomposed into vertices and directed edge pairs is a simplification of the quad-edge case in Guibas & Stolfi (1985) \doi{10.1145/282918.282923}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/wkpool}
  \item Report bugs at \url{https://github.com/hypertidy/wkpool/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
